requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
python-dotenv
streamlit
requests
openai